spectating refactor + small faction role crash fix

+++ server/src/game/contextManager.ts
@@ -8,8 +8,11 @@ import type { Game } from "./game";
 import type { DamageParams } from "./objects/gameObject";
 import type { Player } from "./objects/player";
 enum ContextMode {
+    /** default solos, any map besides factions */
     Solo,
+    /** default duos or squads, any map besides factions */
     Team,
+    /** irrelevant to gamemode type, always the mode if faction map is selected */
     Faction,
 }
 
@@ -68,10 +71,19 @@ export class ContextManager {
     }
 
     isGameStarted(): boolean {
+        // return this._applyContext<boolean>({
+        //     [ContextMode.Solo]: () => this.aliveCount() > 1,
+        //     [ContextMode.Team]: () => this.aliveCount() > 1,
+        //     [ContextMode.Faction]: () => this.aliveCount() > 1,
+        // });
+        return this.aliveCount() > 1;
+    }
+
+    isContextModeSolo(): boolean {
         return this._applyContext<boolean>({
-            [ContextMode.Solo]: () => this.aliveCount() > 1,
-            [ContextMode.Team]: () => this.aliveCount() > 1,
-            [ContextMode.Faction]: () => this.aliveCount() > 1, //tbd
+            [ContextMode.Solo]: () => true,
+            [ContextMode.Team]: () => false,
+            [ContextMode.Faction]: () => false,
         });
     }
 
@@ -103,6 +115,17 @@ export class ContextManager {
                 [...this._game.groups.values()].map((g) => g.livingPlayers),
             [ContextMode.Faction]: () => this._game.teams.map((t) => t.livingPlayers),
         });
+    }
+
+    getSpectatablePlayers(player: Player): Player[] {
+        let playerFilter: (p: Player) => boolean;
+        if (this.getPlayerAlivePlayersContext(player).length != 0) {
+            playerFilter = (p: Player) => !p.disconnected && p.teamId == player.teamId;
+        } else {
+            playerFilter = (p: Player) => !p.disconnected;
+        }
+        //livingPlayers is used here instead of a more "efficient" option because its sorted while other options are not
+        return this._game.playerBarn.livingPlayers.filter(playerFilter);
     }
 
     getPlayerStatusPlayers(player: Player): Player[] | undefined {
@@ -155,13 +178,21 @@ export class ContextManager {
             );
     }
 
-    /** if the current context mode supports reviving, unrelated to individual players */
-    canRevive(): boolean {
-        return this._applyContext<boolean>({
-            [ContextMode.Solo]: () => false,
-            [ContextMode.Team]: () => true,
-            [ContextMode.Faction]: () => true,
-        });
+    /**
+     * by default, true if the current context mode supports reviving
+     *
+     * always true regardless of context mode with self revive
+     * @param playerReviving player that initializes the revive action
+     */
+    canRevive(playerReviving: Player): boolean {
+        return (
+            playerReviving.hasPerk("self_revive") ||
+            this._applyContext<boolean>({
+                [ContextMode.Solo]: () => false,
+                [ContextMode.Team]: () => true,
+                [ContextMode.Faction]: () => true,
+            })
+        );
     }
 
     isReviving(player: Player): boolean {
@@ -288,8 +319,11 @@ export class ContextManager {
 
                 const allDeadOrDisconnected = group.checkAllDeadOrDisconnected(player);
                 const allDowned = group.checkAllDowned(player);
+                const groupHasSelfRevive = group.livingPlayers.find((p) =>
+                    p.hasPerk("self_revive"),
+                );
 
-                if (allDeadOrDisconnected || allDowned) {
+                if (!groupHasSelfRevive && (allDeadOrDisconnected || allDowned)) {
                     group.allDeadOrDisconnected = true; // must set before any kill() calls so the gameovermsgs are accurate
                     player.kill(params);
                     if (allDowned) {
